requirements.txt
ollama
chromadb
mcp
SpeechRecognition
PyAudio
python-dotenv
